build: remote static flag on gcc
This can be passed as CXXFLAGS env var for those who need it. In
majority of cases as is this will cause the build to fail:

	/usr/bin/ld: cannot find -lrt
	/usr/bin/ld: cannot find -lstdc++
	/usr/bin/ld: cannot find -lm
	/usr/bin/ld: cannot find -lc

Also -lrt isn't need in most distros nowadays. A more correct fix would be to
check whether it's needed, but this doesn't seem to be the approach of a
Makefile-only build.